Phase B1: extract HTTP and Prometheus into src/http/
Moves ~530 lines of HTTP router, handlers, version/healthz/Prometheus
rendering structs and functions from main.rs into crates/defra-agent-cli/src/http/
(router.rs, version.rs, healthz.rs, prometheus.rs, mod.rs). main.rs drops
from 7816 → 7284 lines. Zero behavior change; all 68 tests pass.
