+++ crates/defra-agent-cli/src/http/mod.rs
@@ -1,0 +1,6 @@
+pub(crate) mod healthz;
+pub(crate) mod prometheus;
+pub(crate) mod router;
+pub(crate) mod version;
+
+pub(crate) use router::runtime_contract_router;
