+++ crates/defra-agent-cli/src/http/prometheus.rs
@@ -1,0 +1,323 @@
+use anyhow::{Context, Result};
+use serde::{Deserialize, Serialize};
+use serde_json::Value;
+
+use crate::post_graphql;
+
+#[derive(Debug, Deserialize, Serialize)]
+pub(crate) struct MetricsQueryData {
+    #[serde(rename = "AgentRuntime", default)]
+    pub(crate) agent_runtimes: Vec<MetricsRuntimeRow>,
+    #[serde(rename = "InferenceBackend", default)]
+    pub(crate) inference_backends: Vec<MetricsBackendRow>,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub(crate) struct MetricsRuntimeRow {
+    pub(crate) agent_did: String,
+    #[serde(default)]
+    pub(crate) process_state: String,
+    #[serde(default)]
+    pub(crate) reconcile_phase: String,
+    #[serde(default)]
+    pub(crate) active_generation: i64,
+    #[serde(default)]
+    pub(crate) router_generation: i64,
+    #[serde(default)]
+    pub(crate) runnable_behavior_count: i64,
+    #[serde(default)]
+    pub(crate) unavailable_behavior_count: i64,
+    #[serde(default)]
+    pub(crate) last_reconcile_result: String,
+    #[serde(default)]
+    pub(crate) last_reconcile_completed_at: String,
+}
+
+#[derive(Debug, Deserialize, Serialize)]
+pub(crate) struct MetricsBackendRow {
+    pub(crate) backend_id: String,
+    #[serde(default)]
+    pub(crate) enabled: bool,
+    #[serde(default)]
+    pub(crate) max_concurrent: i64,
+    #[serde(default)]
+    pub(crate) max_queue_depth: i64,
+    #[serde(default)]
+    pub(crate) probe_status: String,
+    pub(crate) last_probe: Option<String>,
+}
+
+pub(crate) async fn render_prometheus_metrics(graphql: &str) -> Result<String> {
+    let data = load_metrics_query_data(graphql).await?;
+
+    let mut lines = Vec::new();
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_up",
+        "Whether the defra-agent process is serving.",
+    );
+    push_metric_sample(&mut lines, "defra_agent_up", &[], 1);
+
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_process_state",
+        "One-hot process lifecycle state for each agent runtime.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_reconcile_phase",
+        "One-hot reconcile phase for each agent runtime.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_last_reconcile_result",
+        "One-hot last reconcile result for each agent runtime.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_active_generation",
+        "Current active runtime generation.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_router_generation",
+        "Current router-observed runtime generation.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_runnable_behaviors",
+        "Number of runnable behaviors in the active runtime snapshot.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_unavailable_behaviors",
+        "Number of unavailable behaviors in the active runtime snapshot.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_runtime_last_reconcile_completed_at_seconds",
+        "Unix timestamp of the last completed reconcile.",
+    );
+
+    for runtime in &data.agent_runtimes {
+        let agent_did = runtime.agent_did.clone();
+        for state in [
+            "uninitialized",
+            "recovering",
+            "ready",
+            "shuttingDown",
+            "shutdown",
+        ] {
+            push_metric_sample(
+                &mut lines,
+                "defra_agent_runtime_process_state",
+                &[
+                    ("agent_did", agent_did.clone()),
+                    ("state", state.to_string()),
+                ],
+                i64::from(runtime.process_state == state),
+            );
+        }
+        for phase in ["idle", "debouncing", "resolving", "diffing", "applying"] {
+            push_metric_sample(
+                &mut lines,
+                "defra_agent_runtime_reconcile_phase",
+                &[
+                    ("agent_did", agent_did.clone()),
+                    ("phase", phase.to_string()),
+                ],
+                i64::from(runtime.reconcile_phase == phase),
+            );
+        }
+        for result in ["startup", "noop", "applied", "error"] {
+            push_metric_sample(
+                &mut lines,
+                "defra_agent_runtime_last_reconcile_result",
+                &[
+                    ("agent_did", agent_did.clone()),
+                    ("result", result.to_string()),
+                ],
+                i64::from(runtime.last_reconcile_result == result),
+            );
+        }
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_runtime_active_generation",
+            &[("agent_did", agent_did.clone())],
+            runtime.active_generation,
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_runtime_router_generation",
+            &[("agent_did", agent_did.clone())],
+            runtime.router_generation,
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_runtime_runnable_behaviors",
+            &[("agent_did", agent_did.clone())],
+            runtime.runnable_behavior_count,
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_runtime_unavailable_behaviors",
+            &[("agent_did", agent_did.clone())],
+            runtime.unavailable_behavior_count,
+        );
+        if let Some(timestamp) = rfc3339_timestamp_seconds(&runtime.last_reconcile_completed_at) {
+            push_metric_sample(
+                &mut lines,
+                "defra_agent_runtime_last_reconcile_completed_at_seconds",
+                &[("agent_did", agent_did)],
+                timestamp,
+            );
+        }
+    }
+
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_backend_enabled",
+        "Whether an inference backend is enabled.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_backend_max_concurrent",
+        "Configured maximum concurrency for an inference backend.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_backend_max_queue_depth",
+        "Configured admission queue depth for an inference backend.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_backend_probe_status",
+        "Current probe status for an inference backend.",
+    );
+    push_metric_prelude(
+        &mut lines,
+        "defra_agent_backend_last_probe_seconds",
+        "Unix timestamp of the last backend probe.",
+    );
+
+    for backend in &data.inference_backends {
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_backend_enabled",
+            &[("backend_id", backend.backend_id.clone())],
+            i64::from(backend.enabled),
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_backend_max_concurrent",
+            &[("backend_id", backend.backend_id.clone())],
+            backend.max_concurrent,
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_backend_max_queue_depth",
+            &[("backend_id", backend.backend_id.clone())],
+            backend.max_queue_depth,
+        );
+        push_metric_sample(
+            &mut lines,
+            "defra_agent_backend_probe_status",
+            &[
+                ("backend_id", backend.backend_id.clone()),
+                ("status", backend.probe_status.clone()),
+            ],
+            1,
+        );
+        if let Some(timestamp) = backend
+            .last_probe
+            .as_deref()
+            .and_then(rfc3339_timestamp_seconds)
+        {
+            push_metric_sample(
+                &mut lines,
+                "defra_agent_backend_last_probe_seconds",
+                &[("backend_id", backend.backend_id.clone())],
+                timestamp,
+            );
+        }
+    }
+
+    lines.push(String::new());
+    Ok(lines.join("\n"))
+}
+
+pub(crate) async fn load_metrics_query_data(graphql: &str) -> Result<MetricsQueryData> {
+    let response = post_graphql(
+        graphql,
+        r#"{
+            AgentRuntime {
+                agent_did
+                process_state
+                reconcile_phase
+                active_generation
+                router_generation
+                runnable_behavior_count
+                unavailable_behavior_count
+                last_reconcile_result
+                last_reconcile_completed_at
+            }
+            InferenceBackend {
+                backend_id
+                enabled
+                max_concurrent
+                max_queue_depth
+                probe_status
+                last_probe
+            }
+        }"#,
+    )
+    .await?;
+    let data = response
+        .get("data")
+        .cloned()
+        .unwrap_or_else(|| Value::Object(Default::default()));
+    serde_json::from_value(data).context("decoding runtime HTTP query response")
+}
+
+fn push_metric_prelude(lines: &mut Vec<String>, name: &str, help: &str) {
+    lines.push(format!("# HELP {name} {help}"));
+    lines.push(format!("# TYPE {name} gauge"));
+}
+
+fn push_metric_sample(
+    lines: &mut Vec<String>,
+    name: &str,
+    labels: &[(&str, String)],
+    value: impl std::fmt::Display,
+) {
+    lines.push(format!("{name}{} {value}", format_metric_labels(labels),));
+}
+
+fn format_metric_labels(labels: &[(&str, String)]) -> String {
+    if labels.is_empty() {
+        return String::new();
+    }
+    let rendered = labels
+        .iter()
+        .map(|(key, value)| format!(r#"{key}="{}""#, escape_prometheus_label(value)))
+        .collect::<Vec<_>>()
+        .join(",");
+    format!("{{{rendered}}}")
+}
+
+fn escape_prometheus_label(value: &str) -> String {
+    value
+        .replace('\\', "\\\\")
+        .replace('\n', "\\n")
+        .replace('"', "\\\"")
+}
+
+fn rfc3339_timestamp_seconds(value: &str) -> Option<i64> {
+    let value = value.trim();
+    if value.is_empty() {
+        return None;
+    }
+    chrono::DateTime::parse_from_rfc3339(value)
+        .ok()
+        .map(|timestamp| timestamp.timestamp())
+}
