+++ crates/defra-agent-cli/src/main.rs
@@ -4,16 +4,9 @@ use std::io::{self, BufRead, Read, Write};
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
 use anyhow::{Context, Result};
-use axum::{
-    extract::State,
-    http::{header, StatusCode},
-    response::{IntoResponse, Response},
-    routing::get,
-    Router,
-};
 use clap::{Parser, Subcommand, ValueEnum};
 use defra_agent::defra_node::EmbeddedNode;
 use defra_agent::graphql::escape_graphql_string;
@@ -32,15 +25,16 @@ use serde_json::{json, Map, Value};
 use tokio::sync::watch;
 
 mod desired_state;
+mod http;
 mod telemetry;
+
+use http::runtime_contract_router;
+use http::version::{NodeIdentityResponse, P2pShareableAddressResponse};
 
 const DEFAULT_AGENT_NAME: &str = "default";
 const DEFAULT_INIT_ENDPOINT: &str = "http://localhost:11434/v1";
 const DEFAULT_INIT_MODEL_NAME: &str = "gemma4-26b-a4b";
 const DEFAULT_HTTP_PORT: u16 = 9191;
-const PROMETHEUS_CONTENT_TYPE: &str = "text/plain; version=0.0.4; charset=utf-8";
-const SERVICE_NAME: &str = "defra-agent";
-const SERVICE_BINARY: &str = "defra-agent";
 const DEFAULT_P2P_MAX_CONCURRENT_DAG_FETCHES: usize = 4;
 const DEFAULT_P2P_MAX_CONCURRENT_PUSH_TASKS: usize = 8;
 const DEFAULT_P2P_RATE_LIMIT_BURST: u32 = 500;
@@ -722,37 +716,6 @@ struct StoredRuntimeState {
     p2p_listen_addresses: Vec<String>,
 }
 
-#[derive(Debug, Clone, Serialize)]
-struct VersionResponse {
-    service: &'static str,
-    binary: &'static str,
-    package: &'static str,
-    version: &'static str,
-    repository: &'static str,
-    build: BuildMetadata,
-}
-
-#[derive(Debug, Clone, Serialize)]
-struct BuildMetadata {
-    git_sha: Option<&'static str>,
-    git_ref: Option<&'static str>,
-    git_dirty: Option<bool>,
-    target: Option<&'static str>,
-    profile: Option<&'static str>,
-}
-
-#[derive(Debug, Deserialize)]
-struct NodeIdentityResponse {
-    #[serde(rename = "PeerID")]
-    peer_id: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
-struct P2pShareableAddressResponse {
-    #[serde(default)]
-    address: Option<String>,
-}
-
 #[derive(Debug, Deserialize)]
 struct P2pPeerRow {
     id: String,
@@ -863,56 +826,6 @@ struct ConfigApplyReport {
     planned: desired_state::DesiredStateDiffCollectionsCounts,
     applied: ConfigApplyCounts,
     remaining: desired_state::DesiredStateDiffCollectionsCounts,
-}
-
-#[derive(Clone)]
-struct RuntimeHttpState {
-    graphql: String,
-    started_at: String,
-    started_instant: Instant,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct MetricsQueryData {
-    #[serde(rename = "AgentRuntime", default)]
-    agent_runtimes: Vec<MetricsRuntimeRow>,
-    #[serde(rename = "InferenceBackend", default)]
-    inference_backends: Vec<MetricsBackendRow>,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct MetricsRuntimeRow {
-    agent_did: String,
-    #[serde(default)]
-    process_state: String,
-    #[serde(default)]
-    reconcile_phase: String,
-    #[serde(default)]
-    active_generation: i64,
-    #[serde(default)]
-    router_generation: i64,
-    #[serde(default)]
-    runnable_behavior_count: i64,
-    #[serde(default)]
-    unavailable_behavior_count: i64,
-    #[serde(default)]
-    last_reconcile_result: String,
-    #[serde(default)]
-    last_reconcile_completed_at: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-struct MetricsBackendRow {
-    backend_id: String,
-    #[serde(default)]
-    enabled: bool,
-    #[serde(default)]
-    max_concurrent: i64,
-    #[serde(default)]
-    max_queue_depth: i64,
-    #[serde(default)]
-    probe_status: String,
-    last_probe: Option<String>,
 }
 
 enum ConfigAccess {
@@ -1484,451 +1397,6 @@ fn scheduled_task_row_matches_expected(row: &Value, expected: &Value) -> Result<
         .all(|(key, value)| actual.get(key).is_some_and(|actual| actual == value)))
 }
 
-fn runtime_contract_router(graphql: String) -> Router {
-    let state = RuntimeHttpState {
-        graphql,
-        started_at: chrono::Utc::now().to_rfc3339(),
-        started_instant: Instant::now(),
-    };
-
-    Router::new()
-        .route("/metrics", get(metrics_handler))
-        .route("/version", get(version_handler))
-        .route("/healthz", get(healthz_handler))
-        .with_state(state)
-}
-
-async fn metrics_handler(State(state): State<RuntimeHttpState>) -> Response {
-    match render_prometheus_metrics(&state.graphql).await {
-        Ok(body) => ([(header::CONTENT_TYPE, PROMETHEUS_CONTENT_TYPE)], body).into_response(),
-        Err(error) => (
-            StatusCode::INTERNAL_SERVER_ERROR,
-            [(header::CONTENT_TYPE, "text/plain; charset=utf-8")],
-            format!("metrics render failed: {error}"),
-        )
-            .into_response(),
-    }
-}
-
-async fn version_handler() -> impl IntoResponse {
-    axum::Json(version_response())
-}
-
-async fn healthz_handler(State(state): State<RuntimeHttpState>) -> Response {
-    match load_metrics_query_data(&state.graphql).await {
-        Ok(data) => {
-            let health = render_healthz_payload(&state, Some(&data), None);
-            let status = if health.get("ok") == Some(&Value::Bool(true)) {
-                StatusCode::OK
-            } else {
-                StatusCode::SERVICE_UNAVAILABLE
-            };
-            (status, axum::Json(health)).into_response()
-        }
-        Err(error) => {
-            let health = render_healthz_payload(&state, None, Some(error.to_string()));
-            (StatusCode::SERVICE_UNAVAILABLE, axum::Json(health)).into_response()
-        }
-    }
-}
-
-fn version_response() -> VersionResponse {
-    VersionResponse {
-        service: SERVICE_NAME,
-        binary: SERVICE_BINARY,
-        package: env!("CARGO_PKG_NAME"),
-        version: env!("CARGO_PKG_VERSION"),
-        repository: env!("CARGO_PKG_REPOSITORY"),
-        build: BuildMetadata {
-            git_sha: option_env!("DEFRA_AGENT_BUILD_GIT_SHA"),
-            git_ref: option_env!("DEFRA_AGENT_BUILD_GIT_REF"),
-            git_dirty: option_env!("DEFRA_AGENT_BUILD_GIT_DIRTY").and_then(|value| match value {
-                "true" => Some(true),
-                "false" => Some(false),
-                _ => None,
-            }),
-            target: option_env!("DEFRA_AGENT_BUILD_TARGET"),
-            profile: option_env!("DEFRA_AGENT_BUILD_PROFILE"),
-        },
-    }
-}
-
-fn render_healthz_payload(
-    state: &RuntimeHttpState,
-    data: Option<&MetricsQueryData>,
-    error: Option<String>,
-) -> Value {
-    let version = version_response();
-    let uptime_seconds = state.started_instant.elapsed().as_secs();
-
-    match data {
-        Some(data) => {
-            let runtime_ready = data
-                .agent_runtimes
-                .iter()
-                .any(|runtime| runtime.process_state == ProcessLifecycleState::Ready.as_str());
-            let runtime_degraded = data
-                .agent_runtimes
-                .iter()
-                .any(|runtime| runtime.unavailable_behavior_count > 0);
-            let backend_degraded = data
-                .inference_backends
-                .iter()
-                .any(|backend| backend.enabled && backend.probe_status != "healthy");
-            let ok = runtime_ready;
-            let status = if !runtime_ready {
-                "unhealthy"
-            } else if runtime_degraded || backend_degraded {
-                "degraded"
-            } else {
-                "ok"
-            };
-            let runtime_status = if runtime_ready {
-                if runtime_degraded {
-                    "degraded"
-                } else {
-                    "ok"
-                }
-            } else {
-                "unhealthy"
-            };
-            let backend_status = if backend_degraded { "degraded" } else { "ok" };
-
-            json!({
-                "status": status,
-                "ok": ok,
-                "service": SERVICE_NAME,
-                "version": version.version,
-                "started_at": state.started_at,
-                "uptime_seconds": uptime_seconds,
-                "checks": {
-                    "http": {
-                        "status": "ok",
-                    },
-                    "graphql": {
-                        "status": "ok",
-                        "endpoint": state.graphql,
-                    },
-                    "runtime": {
-                        "status": runtime_status,
-                        "ready": runtime_ready,
-                        "count": data.agent_runtimes.len(),
-                    },
-                    "backends": {
-                        "status": backend_status,
-                        "count": data.inference_backends.len(),
-                    },
-                },
-                "runtimes": data.agent_runtimes,
-                "backends": data.inference_backends,
-            })
-        }
-        None => json!({
-            "status": "unhealthy",
-            "ok": false,
-            "service": SERVICE_NAME,
-            "version": version.version,
-            "started_at": state.started_at,
-            "uptime_seconds": uptime_seconds,
-            "checks": {
-                "http": {
-                    "status": "ok",
-                },
-                "graphql": {
-                    "status": "unhealthy",
-                    "endpoint": state.graphql,
-                    "error": error.unwrap_or_else(|| "runtime GraphQL status unavailable".to_string()),
-                },
-                "runtime": {
-                    "status": "unknown",
-                    "ready": false,
-                    "count": 0,
-                },
-                "backends": {
-                    "status": "unknown",
-                    "count": 0,
-                },
-            },
-            "runtimes": [],
-            "backends": [],
-        }),
-    }
-}
-
-async fn render_prometheus_metrics(graphql: &str) -> Result<String> {
-    let data = load_metrics_query_data(graphql).await?;
-
-    let mut lines = Vec::new();
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_up",
-        "Whether the defra-agent process is serving.",
-    );
-    push_metric_sample(&mut lines, "defra_agent_up", &[], 1);
-
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_process_state",
-        "One-hot process lifecycle state for each agent runtime.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_reconcile_phase",
-        "One-hot reconcile phase for each agent runtime.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_last_reconcile_result",
-        "One-hot last reconcile result for each agent runtime.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_active_generation",
-        "Current active runtime generation.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_router_generation",
-        "Current router-observed runtime generation.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_runnable_behaviors",
-        "Number of runnable behaviors in the active runtime snapshot.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_unavailable_behaviors",
-        "Number of unavailable behaviors in the active runtime snapshot.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_runtime_last_reconcile_completed_at_seconds",
-        "Unix timestamp of the last completed reconcile.",
-    );
-
-    for runtime in &data.agent_runtimes {
-        let agent_did = runtime.agent_did.clone();
-        for state in [
-            "uninitialized",
-            "recovering",
-            "ready",
-            "shuttingDown",
-            "shutdown",
-        ] {
-            push_metric_sample(
-                &mut lines,
-                "defra_agent_runtime_process_state",
-                &[
-                    ("agent_did", agent_did.clone()),
-                    ("state", state.to_string()),
-                ],
-                i64::from(runtime.process_state == state),
-            );
-        }
-        for phase in ["idle", "debouncing", "resolving", "diffing", "applying"] {
-            push_metric_sample(
-                &mut lines,
-                "defra_agent_runtime_reconcile_phase",
-                &[
-                    ("agent_did", agent_did.clone()),
-                    ("phase", phase.to_string()),
-                ],
-                i64::from(runtime.reconcile_phase == phase),
-            );
-        }
-        for result in ["startup", "noop", "applied", "error"] {
-            push_metric_sample(
-                &mut lines,
-                "defra_agent_runtime_last_reconcile_result",
-                &[
-                    ("agent_did", agent_did.clone()),
-                    ("result", result.to_string()),
-                ],
-                i64::from(runtime.last_reconcile_result == result),
-            );
-        }
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_runtime_active_generation",
-            &[("agent_did", agent_did.clone())],
-            runtime.active_generation,
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_runtime_router_generation",
-            &[("agent_did", agent_did.clone())],
-            runtime.router_generation,
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_runtime_runnable_behaviors",
-            &[("agent_did", agent_did.clone())],
-            runtime.runnable_behavior_count,
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_runtime_unavailable_behaviors",
-            &[("agent_did", agent_did.clone())],
-            runtime.unavailable_behavior_count,
-        );
-        if let Some(timestamp) = rfc3339_timestamp_seconds(&runtime.last_reconcile_completed_at) {
-            push_metric_sample(
-                &mut lines,
-                "defra_agent_runtime_last_reconcile_completed_at_seconds",
-                &[("agent_did", agent_did)],
-                timestamp,
-            );
-        }
-    }
-
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_backend_enabled",
-        "Whether an inference backend is enabled.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_backend_max_concurrent",
-        "Configured maximum concurrency for an inference backend.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_backend_max_queue_depth",
-        "Configured admission queue depth for an inference backend.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_backend_probe_status",
-        "Current probe status for an inference backend.",
-    );
-    push_metric_prelude(
-        &mut lines,
-        "defra_agent_backend_last_probe_seconds",
-        "Unix timestamp of the last backend probe.",
-    );
-
-    for backend in &data.inference_backends {
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_backend_enabled",
-            &[("backend_id", backend.backend_id.clone())],
-            i64::from(backend.enabled),
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_backend_max_concurrent",
-            &[("backend_id", backend.backend_id.clone())],
-            backend.max_concurrent,
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_backend_max_queue_depth",
-            &[("backend_id", backend.backend_id.clone())],
-            backend.max_queue_depth,
-        );
-        push_metric_sample(
-            &mut lines,
-            "defra_agent_backend_probe_status",
-            &[
-                ("backend_id", backend.backend_id.clone()),
-                ("status", backend.probe_status.clone()),
-            ],
-            1,
-        );
-        if let Some(timestamp) = backend
-            .last_probe
-            .as_deref()
-            .and_then(rfc3339_timestamp_seconds)
-        {
-            push_metric_sample(
-                &mut lines,
-                "defra_agent_backend_last_probe_seconds",
-                &[("backend_id", backend.backend_id.clone())],
-                timestamp,
-            );
-        }
-    }
-
-    lines.push(String::new());
-    Ok(lines.join("\n"))
-}
-
-async fn load_metrics_query_data(graphql: &str) -> Result<MetricsQueryData> {
-    let response = post_graphql(
-        graphql,
-        r#"{
-            AgentRuntime {
-                agent_did
-                process_state
-                reconcile_phase
-                active_generation
-                router_generation
-                runnable_behavior_count
-                unavailable_behavior_count
-                last_reconcile_result
-                last_reconcile_completed_at
-            }
-            InferenceBackend {
-                backend_id
-                enabled
-                max_concurrent
-                max_queue_depth
-                probe_status
-                last_probe
-            }
-        }"#,
-    )
-    .await?;
-    let data = response
-        .get("data")
-        .cloned()
-        .unwrap_or_else(|| Value::Object(Default::default()));
-    serde_json::from_value(data).context("decoding runtime HTTP query response")
-}
-
-fn push_metric_prelude(lines: &mut Vec<String>, name: &str, help: &str) {
-    lines.push(format!("# HELP {name} {help}"));
-    lines.push(format!("# TYPE {name} gauge"));
-}
-
-fn push_metric_sample(
-    lines: &mut Vec<String>,
-    name: &str,
-    labels: &[(&str, String)],
-    value: impl std::fmt::Display,
-) {
-    lines.push(format!("{name}{} {value}", format_metric_labels(labels),));
-}
-
-fn format_metric_labels(labels: &[(&str, String)]) -> String {
-    if labels.is_empty() {
-        return String::new();
-    }
-    let rendered = labels
-        .iter()
-        .map(|(key, value)| format!(r#"{key}="{}""#, escape_prometheus_label(value)))
-        .collect::<Vec<_>>()
-        .join(",");
-    format!("{{{rendered}}}")
-}
-
-fn escape_prometheus_label(value: &str) -> String {
-    value
-        .replace('\\', "\\\\")
-        .replace('\n', "\\n")
-        .replace('"', "\\\"")
-}
-
-fn rfc3339_timestamp_seconds(value: &str) -> Option<i64> {
-    let value = value.trim();
-    if value.is_empty() {
-        return None;
-    }
-    chrono::DateTime::parse_from_rfc3339(value)
-        .ok()
-        .map(|timestamp| timestamp.timestamp())
-}
 
 #[derive(Subcommand)]
 enum BackendCommand {
